Added links

+++ Control/HomeControl.php
@@ -6,40 +6,43 @@ include_once dirname(__FILE__) . '/../Model/Database/EditableAreaDAO.php';
 include_once dirname(__FILE__) . '/../Model/Database/CommentDAO.php';
 
 class HomeControl extends Control {
-    
-    function __construct($model) {
-        parent::__construct($model);
-    }
 
-    
-    public function initialize() {
-        $type = EditableArea::NEWS;
-        if(isset($_POST['deleteNews'])){
-            $id = (int)sanitizeString($_POST['id']);
-            $this->delete($id);
-        }
-        $this->model->news = EditableAreaDAO::selectByEditableAreaType($type);
-    }
+	function __construct($model) {
+		parent::__construct($model);
+	}
 
-    private function delete($id){
-        if(EditableAreaDAO::editableAreaExists($id)){
-            $news = EditableAreaDAO::selectById($id);
-            $comments = CommentDAO::selectByTypeAndTarget(Comment::NEWS, $id);
-            foreach($comments as $comment){
-                $commentId = (int)$comment->getIdComment();
-                $this->deleteComment($commentId);
-            }
-            EditableAreaDAO::delete($news);
-        }
-    }
+	public function initialize() {
+		global $loggedIn;
+		global $admin;
+		$type = EditableArea::NEWS;
+		if($admin && $loggedIn) {
+			if(isset($_POST['deleteNews'])) {
+				$id = (int)sanitizeString($_POST['id']);
+				$this->delete($id);
+			}
+		}
+		$this->model->news = EditableAreaDAO::selectByEditableAreaType($type);
+	}
 
-    private function deleteComment($id) {
-        $comments = CommentDAO::selectByTypeAndTarget(Comment::RE, $id);
-        foreach ($comments as $comment) {
-            CommentDAO::delete($comment);
-        }
-        $news = CommentDAO::selectById($id);
-        CommentDAO::delete($news);
-    }
+	private function delete($id) {
+		if(EditableAreaDAO::editableAreaExists($id)) {
+			$news = EditableAreaDAO::selectById($id);
+			$comments = CommentDAO::selectByTypeAndTarget(Comment::NEWS, $id);
+			foreach ($comments as $comment) {
+				$commentId = (int)$comment->getIdComment();
+				$this->deleteComment($commentId);
+			}
+			EditableAreaDAO::delete($news);
+		}
+	}
+
+	private function deleteComment($id) {
+		$comments = CommentDAO::selectByTypeAndTarget(Comment::RE, $id);
+		foreach ($comments as $comment) {
+			CommentDAO::delete($comment);
+		}
+		$news = CommentDAO::selectById($id);
+		CommentDAO::delete($news);
+	}
 
 }
